+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added ability to download image files generated with `generateImage` client-side without sending 
   data to the server.
 * Used the newly added `generateImage` and `imageData` properties to enable svg generation using https://github.com/kinimesi/cytoscape-svg
+* Added responsive cytoscape.js graph feature toggled using the `responsive` property.
+* One new demo:
+    * `demos/usage-responsive-graph.py`: Example of graph with the ability to toggle the responsive feature on and off.
 
 ### Changed
 * `utils.Tree`: v0.1.1 broke compatibility with Python 2. Therefore, modified code to be compatible